+++ pyosu/http.py
@@ -28,23 +28,32 @@ from .exceptions import WrongApiKey, RouteNotFound, InvalidArgument, HTTPError, 
 
 class Route:
 
-    BASE = 'https://osu.ppy.sh/api/'
+    def __init__(self, path : str = '', api_key : str = '', base = 'https://osu.ppy.sh/api/', **parameters):
 
-    def __init__(self, path : str, api_key : str, **parameters):
-
+        self.base        = base
         self.path        = path
         self.api_key     = api_key
-        self.parameters  = parameters
+        self.parameters  = {}
+
+        for key, value in parameters.items():
+            if value != None:
+                self.parameters[key] = value
 
     @property
     def route(self):
         """ Returns the current route """
 
-        params = ''
-        for key, value in self.parameters.items():
-            params += f'&{str(key)}={str(value)}'
+        params = []
+        if self.api_key != '':
+            params.append(f'k={self.api_key}')
 
-        return f"{Route.BASE}{self.path}?k={self.api_key}{params}"
+        for key, value in self.parameters.items():
+            params.append(f'{str(key)}={str(value)}')
+
+        if len(params) > 0:
+            return f"{self.base}{self.path}?{'&'.join(params)}"
+
+        return f"{self.base}{self.path}"
         
     def add_param(self, key, value):
         """ Adds or updates a prameter """
@@ -76,27 +85,14 @@ class Route:
 
         return
 
-    def check_path(self):
-        """ Raise the RouteNotFound Exception if the path isn't in the following list :
-
-            [get_beatmaps, get_user, get_scores, get_user_best, get_user_recent, get_match, get_replay]
-        """
-
-        accepted_paths = ['get_beatmaps', 'get_user', 'get_scores',
-            'get_user_best', 'get_user_recent', 'get_match', 'get_replay']
-
-        if self.path not in accepted_paths:
-            raise RouteNotFound(f'The route {self.route} does not exists.', 404)
-        
-        return
-
 class Request():
 
-    def __init__(self, route : Route, retry : int = 5):
+    def __init__(self, route : Route, retry : int = 5, json_response : bool = True):
 
-        self.retry_count = retry # Number of retrys to do before throwing any errors
-        self.route = route
-        self._data = []
+        self.json_response = json_response
+        self.retry_count   = retry # Number of retrys to do before throwing any errors
+        self.route         = route
+        self._data         = []
 
     @property
     def data(self):
@@ -105,6 +101,9 @@ class Request():
     async def get_json(self, response, *args):
         """ Returns the json version of an api response """
         text = await response.text()
+
+        if self.json_response is False:
+            return text
 
         if (text == None or text == ''):
             return {}
@@ -126,7 +125,6 @@ class Request():
         """ Fetches some data with a session using the actual route """
 
         self.route.check_params()
-        self.route.check_path()
 
         async with session.get(self.route.route) as response: